Struct csplit::Filer

struct Filer<'a> {
    silent: bool,
    prefix: &'a str,
    digits: i32,
    file_counter: i32,
    created: &'a mut Vec<String>,
    writer: File,
    bytes: usize,
    current: String,
    pub buffer: Vec<String>,
}

Output files handler.

The filer will rotate the files on request, write lines to the current file, and keep track of created output files.

Fields
silent: boolprefix: &'a strdigits: i32file_counter: i32created: &'a mut Vec<String>writer: Filebytes: usizecurrent: Stringbuffer: Vec<String>
Methods

impl<'a> Filer<'a>

fn new(
prefix: &'a str,
digits: i32,
created: &'a mut Vec<String>,
silent: bool
) -> Result<Filer<'a>, Error>

fn rotate(&mut self) -> Result<(), Error>

fn write_line(&mut self, line: &str) -> Result<(), Error>

fn flush(&mut self) -> Result<(), Error>