Trait csplit::Matcher

trait Matcher: Debug {
    fn match_line(&mut self, lineno: usize, line: &str) -> (bool, bool);

    fn process_line(
        &mut self, 
        filer: &mut Filer, 
        lineno: usize, 
        line: &str
    ) -> Result<(), Error>;
}

Matcher trait used to match lines.

Required methods

fn match_line(&mut self, lineno: usize, line: &str) -> (bool, bool)

Check if a line matches.

Return a pair of booleans:

First boolean is true if the file rotation should take place.

Second boolean is true if the line matches.

fn process_line(
&mut self,
filer: &mut Filer,
lineno: usize,
line: &str
) -> Result<(), Error>

Process an input line.

Return true if the matcher matched the line, false otherwise.

Loading content...